Fix logging config leak between tests
(Assisted by Claude Code; any errors are mine.)
